+++ src/drawer/index.jsx
@@ -1,0 +1,4 @@
+import ConfigProvider from '../config-provider';
+import Drawer from './drawer';
+
+export default ConfigProvider.config(Drawer);