+++ src/menu/view/sub-menu.jsx
@@ -87,6 +87,19 @@ export default class SubMenu extends Component {
         return openKeys.indexOf(_key) > -1;
     }
 
+    getChildSelected() {
+        const { _key, root } = this.props;
+        const { selectMode } = root.props;
+        const { selectedKeys } = root.state;
+
+        const _keyPos = root.k2n[_key].pos;
+
+        return (
+            !!selectMode &&
+            selectedKeys.some(key => root.k2n[key].pos.indexOf(_keyPos) > -1)
+        );
+    }
+
     handleMouseEnter(e) {
         this.handleOpen(true);
 
@@ -119,6 +132,11 @@ export default class SubMenu extends Component {
         const { mode, root } = this.props;
 
         return Children.map(children, child => {
+            // to fix https://github.com/alibaba-fusion/next/issues/952
+            if (typeof child !== 'function' && typeof child !== 'object') {
+                return child;
+            }
+
             return cloneElement(child, {
                 parent: this,
                 parentMode: mode || root.props.mode,
@@ -150,6 +168,8 @@ export default class SubMenu extends Component {
         } = root.props;
         const triggerType = propsTriggerType || rootTriggerType;
         const open = this.getOpen();
+        const isChildSelected = this.getChildSelected();
+
         const others = obj.pickOthers(
             Object.keys(SubMenu.propTypes),
             this.props
@@ -169,6 +189,10 @@ export default class SubMenu extends Component {
             type: 'submenu',
             component: 'div',
             parentMode,
+            className: cx({
+                [`${prefix}opened`]: open,
+                [`${prefix}child-selected`]: isChildSelected,
+            }),
         };
         const arrorProps = {
             type:
@@ -193,9 +217,6 @@ export default class SubMenu extends Component {
             arrorProps.onClick = this.handleClick;
         } else {
             itemProps.onClick = this.handleClick;
-        }
-        if (open) {
-            itemProps.className = `${prefix}opened`;
         }
 
         const newSubMenuContentClassName = cx({