+++ types/menu-button/index.d.ts
@@ -2,13 +2,14 @@
 
 import * as React from 'react';
 import { Item, Group, Divider } from '../menu';
+import CommonProps from '../util';
 import { ButtonProps } from '../button';
 
-interface HTMLAttributesWeak extends React.HTMLAttributes<HTMLElement> {
+interface HTMLAttributesWeak extends ButtonProps {
     onSelect?: any;
 }
 
-export interface MenuButtonProps extends HTMLAttributesWeak, ButtonProps {
+export interface MenuButtonProps extends HTMLAttributesWeak, CommonProps {
     /**
      * 按钮上的文本内容
      */